+++ modules/apps/site/site-admin-web/src/com/liferay/site/admin/web/customattributes/SiteCustomAttributesDisplay.java
@@ -1,0 +1,41 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.site.admin.web.customattributes;
+
+import com.liferay.portal.model.Group;
+import com.liferay.portal.theme.ThemeDisplay;
+import com.liferay.portlet.expando.model.BaseCustomAttributesDisplay;
+
+/**
+ * @author Jorge Ferrer
+ */
+public class SiteCustomAttributesDisplay extends BaseCustomAttributesDisplay {
+
+	@Override
+	public String getClassName() {
+		return Group.class.getName();
+	}
+
+	@Override
+	public String getIconCssClass() {
+		return "icon-globe";
+	}
+
+	@Override
+	public String getIconPath(ThemeDisplay themeDisplay) {
+		return themeDisplay.getPathThemeImages() + "/common/site_icon.png";
+	}
+
+}
