+++ modules/apps/site/site-admin-web/src/com/liferay/site/admin/web/taglib/ui/SiteContentSharingFormNavigatorEntry.java
@@ -12,11 +12,13 @@
  * details.
  */
 
-package com.liferay.portal.servlet.taglib.ui;
+package com.liferay.site.admin.web.taglib.ui;
 
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorConstants;
 import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
+import com.liferay.portal.kernel.util.PrefsPropsUtil;
+import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.model.User;
 
@@ -26,27 +28,35 @@ import java.util.Locale;
  * @author Sergio González
  */
 @OSGiBeanProperties(property = {"service.ranking:Integer=20"})
-public class SiteDocumentsAndMediaFormNavigatorEntry
+public class SiteContentSharingFormNavigatorEntry
 	extends BaseSiteFormNavigatorEntry {
 
 	@Override
 	public String getCategoryKey() {
-		return FormNavigatorConstants.CATEGORY_KEY_SITES_BASIC_INFORMATION;
+		return FormNavigatorConstants.CATEGORY_KEY_SITES_ADVANCED;
 	}
 
 	@Override
 	public String getKey() {
-		return "documents-and-media";
+		return "content-sharing";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "documents-and-media");
+		return LanguageUtil.get(locale, "content-sharing");
 	}
 
 	@Override
 	public boolean isVisible(User user, Group group) {
-		if ((group == null) || group.isCompany()) {
+		if (group == null) {
+			return false;
+		}
+
+		int contentSharingWithChildrenEnabled = PrefsPropsUtil.getInteger(
+			group.getCompanyId(),
+			PropsKeys.SITES_CONTENT_SHARING_WITH_CHILDREN_ENABLED);
+
+		if (contentSharingWithChildrenEnabled == 0) {
 			return false;
 		}
 
@@ -55,7 +65,7 @@ public class SiteDocumentsAndMediaFormNavigatorEntry
 
 	@Override
 	protected String getJspPath() {
-		return "/html/portlet/sites_admin/site/documents_and_media.jsp";
+		return "/site/content_sharing.jsp";
 	}
 
 }
