+++ modules/apps/site/site-admin-web/src/com/liferay/site/admin/web/taglib/ui/SiteDocumentsAndMediaFormNavigatorEntry.java
@@ -12,38 +12,36 @@
  * details.
  */
 
-package com.liferay.portal.servlet.taglib.ui;
+package com.liferay.site.admin.web.taglib.ui;
 
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorConstants;
 import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.model.User;
-import com.liferay.portlet.ratings.definition.PortletRatingsDefinitionUtil;
-import com.liferay.portlet.ratings.definition.PortletRatingsDefinitionValues;
 
 import java.util.Locale;
-import java.util.Map;
 
 /**
  * @author Sergio González
  */
-@OSGiBeanProperties(property = {"service.ranking:Integer=10"})
-public class SiteRatingsFormNavigatorEntry extends BaseSiteFormNavigatorEntry {
+@OSGiBeanProperties(property = {"service.ranking:Integer=20"})
+public class SiteDocumentsAndMediaFormNavigatorEntry
+	extends BaseSiteFormNavigatorEntry {
 
 	@Override
 	public String getCategoryKey() {
-		return FormNavigatorConstants.CATEGORY_KEY_SITES_MISCELLANEOUS;
+		return FormNavigatorConstants.CATEGORY_KEY_SITES_BASIC_INFORMATION;
 	}
 
 	@Override
 	public String getKey() {
-		return "ratings";
+		return "documents-and-media";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "ratings");
+		return LanguageUtil.get(locale, "documents-and-media");
 	}
 
 	@Override
@@ -52,17 +50,12 @@ public class SiteRatingsFormNavigatorEntry extends BaseSiteFormNavigatorEntry {
 			return false;
 		}
 
-		Map<String, PortletRatingsDefinitionValues>
-			portletRatingsDefinitionValuesMap =
-				PortletRatingsDefinitionUtil.
-					getPortletRatingsDefinitionValuesMap();
-
-		return !portletRatingsDefinitionValuesMap.isEmpty();
+		return true;
 	}
 
 	@Override
 	protected String getJspPath() {
-		return "/html/portlet/sites_admin/site/ratings.jsp";
+		return "/site/documents_and_media.jsp";
 	}
 
 }
