+++ modules/apps/site/site-admin-web/src/com/liferay/site/admin/web/taglib/ui/SiteStagingFormNavigatorEntry.java
@@ -12,11 +12,14 @@
  * details.
  */
 
-package com.liferay.portal.servlet.taglib.ui;
+package com.liferay.site.admin.web.taglib.ui;
 
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorConstants;
 import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
+import com.liferay.portal.model.Group;
+import com.liferay.portal.model.User;
+import com.liferay.portal.util.PropsValues;
 
 import java.util.Locale;
 
@@ -24,26 +27,42 @@ import java.util.Locale;
  * @author Sergio González
  */
 @OSGiBeanProperties(property = {"service.ranking:Integer=50"})
-public class SiteDetailsFormNavigatorEntry extends BaseSiteFormNavigatorEntry {
+public class SiteStagingFormNavigatorEntry extends BaseSiteFormNavigatorEntry {
 
 	@Override
 	public String getCategoryKey() {
-		return FormNavigatorConstants.CATEGORY_KEY_SITES_BASIC_INFORMATION;
+		return FormNavigatorConstants.CATEGORY_KEY_SITES_ADVANCED;
 	}
 
 	@Override
 	public String getKey() {
-		return "details";
+		return "staging";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "details");
+		return LanguageUtil.get(locale, "staging");
+	}
+
+	@Override
+	public boolean isVisible(User user, Group group) {
+		if (group == null) {
+			return false;
+		}
+
+		if (group.hasStagingGroup() ||
+			(group.hasRemoteStagingGroup() &&
+			 !PropsValues.STAGING_LIVE_GROUP_REMOTE_STAGING_ENABLED)) {
+
+			return false;
+		}
+
+		return true;
 	}
 
 	@Override
 	protected String getJspPath() {
-		return "/html/portlet/sites_admin/site/details.jsp";
+		return "/site/staging.jsp";
 	}
 
 }
