+++ modules/apps/site/site-admin-web/src/com/liferay/site/admin/web/taglib/ui/SitesMiscellaneousFormNavigatorCategory.java
@@ -12,15 +12,12 @@
  * details.
  */
 
-package com.liferay.portal.servlet.taglib.ui;
+package com.liferay.site.admin.web.taglib.ui;
 
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorCategory;
 import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorConstants;
 import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
-import com.liferay.portal.kernel.util.PrefsPropsUtil;
-import com.liferay.portal.kernel.util.PropsKeys;
-import com.liferay.portal.model.Group;
-import com.liferay.portal.model.User;
 
 import java.util.Locale;
 
@@ -28,43 +25,22 @@ import java.util.Locale;
  * @author Sergio González
  */
 @OSGiBeanProperties(property = {"service.ranking:Integer=10"})
-public class SiteRecycleBinFormNavigatorEntry
-	extends BaseSiteFormNavigatorEntry {
+public class SitesMiscellaneousFormNavigatorCategory
+	implements FormNavigatorCategory {
 
 	@Override
-	public String getCategoryKey() {
-		return FormNavigatorConstants.CATEGORY_KEY_SITES_ADVANCED;
+	public String getFormNavigatorId() {
+		return FormNavigatorConstants.FORM_NAVIGATOR_ID_SITES;
 	}
 
 	@Override
 	public String getKey() {
-		return "recycle-bin";
+		return FormNavigatorConstants.CATEGORY_KEY_SITES_MISCELLANEOUS;
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "recycle-bin");
-	}
-
-	@Override
-	public boolean isVisible(User user, Group group) {
-		if (group == null) {
-			return false;
-		}
-
-		boolean trashEnabled = PrefsPropsUtil.getBoolean(
-			group.getCompanyId(), PropsKeys.TRASH_ENABLED);
-
-		if (!trashEnabled) {
-			return false;
-		}
-
-		return true;
-	}
-
-	@Override
-	protected String getJspPath() {
-		return "/html/portlet/sites_admin/site/recycle_bin.jsp";
+		return LanguageUtil.get(locale, "miscellaneous");
 	}
 
 }
